pyproject.toml
[tool.poetry]
name = "discourtesy"
version = "0.1.0"
description = "A minimal framework to handle Discord interactions."
license = "MIT"
authors = ["Robin Mahieu <robin.mahieu@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.5.3"
PyNaCl = "^1.4.0"
starlette = "^0.16.0"
uvicorn = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
